+++ CHANGELOG.md
@@ -4,9 +4,13 @@
 
 ## 2.0
 
+### 2.0.0-beta.3 (2023-03-06)
+
+- New component: LabelWithIndicator [#608](https://github.com/CartoDB/carto-react/pull/608)
+- Design review of Chip component [#607](https://github.com/CartoDB/carto-react/pull/607)
+
 ### 2.0.0-beta.2 (2023-02-17)
 
-- Design review of Chip component [#607](https://github.com/CartoDB/carto-react/pull/607)
 - Fix overwrite default client in SQL post request [#603](https://github.com/CartoDB/carto-react/pull/603)
 
 ### 2.0.0-beta.1 (2023-02-14)